+++ forge16/src/main/java/com/vecoo/extralib/permission/UtilPermission.java
@@ -8,7 +8,7 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraftforge.server.permission.PermissionAPI;
 import net.minecraftforge.server.permission.context.Context;
 
-import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 public class UtilPermission {
@@ -43,7 +43,7 @@ public class UtilPermission {
         return false;
     }
 
-    public static int minValue(int value, PlayerEntity player, List<String> permissionList) {
+    public static int minValue(int value, PlayerEntity player, Set<String> permissionList) {
         for (String permission : permissionList) {
             if (UtilPermission.hasPermission(player, permission)) {
                 value = Math.min(value, Integer.parseInt(permission.substring(permission.lastIndexOf('.') + 1)));
@@ -52,7 +52,7 @@ public class UtilPermission {
         return value;
     }
 
-    public static int maxValue(int value, PlayerEntity player, List<String> permissionList) {
+    public static int maxValue(int value, PlayerEntity player, Set<String> permissionList) {
         for (String permission : permissionList) {
             if (UtilPermission.hasPermission(player, permission)) {
                 value = Math.max(value, Integer.parseInt(permission.substring(permission.lastIndexOf('.') + 1)));
@@ -61,7 +61,7 @@ public class UtilPermission {
         return value;
     }
 
-    public static int minValue(int value, UUID playerUUID, String playerName, List<String> permissionList) {
+    public static int minValue(int value, UUID playerUUID, String playerName, Set<String> permissionList) {
         for (String permission : permissionList) {
             if (UtilPermission.hasPermission(playerUUID, playerName, permission)) {
                 value = Math.min(value, Integer.parseInt(permission.substring(permission.lastIndexOf('.') + 1)));
@@ -70,7 +70,7 @@ public class UtilPermission {
         return value;
     }
 
-    public static int maxValue(int value, UUID playerUUID, String playerName, List<String> permissionList) {
+    public static int maxValue(int value, UUID playerUUID, String playerName, Set<String> permissionList) {
         for (String permission : permissionList) {
             if (UtilPermission.hasPermission(playerUUID, playerName, permission)) {
                 value = Math.max(value, Integer.parseInt(permission.substring(permission.lastIndexOf('.') + 1)));