+++ changelog.md
@@ -1,1 +1,2 @@
-* Added custom storage interface.
+* Unsupported 1.19.2 and Fabric 1.20.1.
+* Minor changes.